NullPointerException handling, connectObexSession
In connectObexSession, if a null mSocket is passed to
BluetoothObexTransport, then a NullPointerException will be thrown when
a new ClientSession is created.

This can occur if abort() --> closeSocket() is called between the calls
to connectSocket() and connectObexSession().

Bug: 142056980
Test: cycle Bluetooth adapter on/off 1k times
CRs-Fixed: 3619240
Change-Id: I78691eb03cd4f3e12922a31b1bc4f4144353c125
(cherry picked from commit 55dd9d6c4cb5389c8fbd9385e9aebd6e01585ae5)