HFP: Cache call state for future when LE-A media is streaming
Use-case: Dut connects two headsets. Active media device is LE-A only
connected headset, active voice device is legacy headset.

Reproduce steps:
1. Play music over LE-A only HS with BAP profile.
2. Start a call. Legacy HS creates eSCO failed.

Root-cause: In BREDR solution, it will cache the call state for future
if current a2dp is PLAYING. But it doesn't include the LE-A media streaming
state. This timing issue happens when create eSCO in process of suspending
LE-A music stream.

Fixed: Add the hasp map mBapConnState to maintance the LE-A media stream state.
Caching the call state for future when LE-A media is PLAYING.

Change-Id: I4b6ff3de26c2b73ffdad36894d2b809eb8b2e527
CRs-Fixed: 3265912