Fix for ActiveDevice discrepancy
Incorrect active device maintained in ActiveDeviceManager and
APM for CISP devices. When second earbud connects, setActive
is called for the second earbud and same is set as
mLeAudioActiveDevice. For group devices first connected earbud is
is considered as active absolute device and setActive for second
earbud is ignored. To avoid discrepancy maintain active absolute
device and mLeAudioActiveDevice same.

Change-Id: I3996d24dcef5c1611e5ccf4eec854dbeb3c7b648