Correct return of BluteoothDevice#createBond
When two apps try to create bond to the same
remote device, the second call would return false,
even when it didn't fail.
This is not consistent with the Javadoc, which states
that the method should return true if bonding is began.

This change makes the method return true when it's already bonding,
which allows callers to expect a BOND_STATE_CHANGED intent
to be broadcast when the bond is complete.

Bug: 269056263
Test: atest BluetoothInstrumentationTests

CRs-Fixed: 3531619
Change-Id: Ic30bb010c09738eecfc8a0285accb4db0ddeeefc
(cherry picked from commit a85acd545274792104b4573a560f673375c05dcf)