btService: get absolute address for getActiveDevice
When a CSIP device is connected, and an external app
queries for activeDevice from Adapter, group address
is returned. This is leading to wrong active device
being shown on external app UI and active device switch
is also not happening.

In the fix, returning absolute address from APM isntead of group
address.

CRs-Fixed: 3526741

Change-Id: I17b278f05bcb7e3e109e1d23f1f3eb7f9e2b5f77