HFP: Remove pending connection attempts that were queue for retry if connected
We have a logic to queue connection messages if the connection
fails to establish.
Remove pending queued messages from the queue if the device connection is initiated by the user or if it is connected as part of an incoming connection attempt from the remote side.
This is to prevent auto connect attempts.

Change-Id: I5332ae820452be211d8265934126e79ce5c563ca
CRs-Fixed: 3920516